Continued WIP on standardizing feature classes (polygon)
